TransactionBuilder and ActionBuilder classes
